ruleset + image download
Updated ruleset.json and image download functions. Firehandler, will now download the wallpapers to the Firebase folder inside the Ambience, and then put the correct file paths in the downloaded JSON.

This means, that all the wallpapers are kept on file yes. This was done to ensure compatibility with what other team members are working on.